Moved code related to job executions
