+++ scale/job/execution/tasks/job_task.py
@@ -1,7 +1,7 @@
 """Defines the class for a job execution job task"""
 from __future__ import unicode_literals
 
-from job.execution.running.tasks.exe_task import JobExecutionTask
+from job.execution.tasks.exe_task import JobExecutionTask
 from job.resources import NodeResources
 
 
@@ -32,7 +32,7 @@ class JobTask(JobExecutionTask):
         self._command_arguments = job_exe.command_arguments
 
     def determine_error(self, task_update):
-        """See :meth:`job.execution.running.tasks.exe_task.JobExecutionTask.determine_error`
+        """See :meth:`job.execution.tasks.exe_task.JobExecutionTask.determine_error`
         """
 
         with self._lock:
@@ -50,7 +50,7 @@ class JobTask(JobExecutionTask):
             return error
 
     def get_resources(self):
-        """See :meth:`job.execution.running.tasks.base_task.Task.get_resources`
+        """See :meth:`job.tasks.base_task.Task.get_resources`
         """
 
         with self._lock:
@@ -58,7 +58,7 @@ class JobTask(JobExecutionTask):
             return NodeResources(cpus=self._cpus, mem=self._mem, disk=self._disk_out)
 
     def populate_job_exe_model(self, job_exe):
-        """See :meth:`job.execution.running.tasks.base_task.Task.populate_job_exe_model`
+        """See :meth:`job.execution.tasks.exe_task.JobExecutionTask.populate_job_exe_model`
         """
 
         with self._lock:
