+++ scale/job/execution/tasks/post_task.py
@@ -1,7 +1,7 @@
 """Defines the class for a job execution post-task"""
 from __future__ import unicode_literals
 
-from job.execution.running.tasks.exe_task import JobExecutionTask
+from job.execution.tasks.exe_task import JobExecutionTask
 from job.management.commands.scale_post_steps import EXIT_CODE_DICT as POST_EXIT_CODE_DICT
 from job.resources import NodeResources
 
@@ -27,7 +27,7 @@ class PostTask(JobExecutionTask):
         self._command_arguments = 'scale_post_steps -i %i' % job_exe.id
 
     def determine_error(self, task_update):
-        """See :meth:`job.execution.running.tasks.exe_task.JobExecutionTask.determine_error`
+        """See :meth:`job.execution.tasks.exe_task.JobExecutionTask.determine_error`
         """
 
         with self._lock:
@@ -45,7 +45,7 @@ class PostTask(JobExecutionTask):
             return error
 
     def get_resources(self):
-        """See :meth:`job.execution.running.tasks.base_task.Task.get_resources`
+        """See :meth:`job.tasks.base_task.Task.get_resources`
         """
 
         with self._lock:
@@ -53,7 +53,7 @@ class PostTask(JobExecutionTask):
             return NodeResources(cpus=self._cpus, mem=self._mem)
 
     def populate_job_exe_model(self, job_exe):
-        """See :meth:`job.execution.running.tasks.base_task.Task.populate_job_exe_model`
+        """See :meth:`job.execution.tasks.exe_task.JobExecutionTask.populate_job_exe_model`
         """
 
         with self._lock:
