+++ scale/scheduler/offer/manager.py
@@ -92,7 +92,7 @@ class OfferManager(object):
         resources
 
         :param job_exe: The running job execution
-        :type job_exe: :class:`job.execution.running.job_exe.RunningJobExecution`
+        :type job_exe: :class:`job.execution.job_exe.RunningJobExecution`
         :returns: One of the OfferManager constants indicating if the next task was accepted or why it was not accepted
         :rtype: int
         """
@@ -108,7 +108,7 @@ class OfferManager(object):
         """Considers if the given task can be run with the current offered resources
 
         :param task: The task to consider
-        :type task: :class:`job.execution.running.tasks.base_task.Task`
+        :type task: :class:`job.tasks.base_task.Task`
         :returns: One of the OfferManager constants indicating if the task was accepted or why it was not accepted
         :rtype: int
         """
